Ensure container's image get updated if their name stay the same.
In some environment the containers' tag isn't modified (think
"latest"). This is already supported during deployment so we re-use
the code during update.

We also make sure that the same mechanism is available for
ovn_controller update. It can be selectively disable if one pass:

    --skip-tags ovn_image

to the external-update run command.

We remove the ignore error as it's wrong indication that we are not
taking errors into account in those tasks. We also adjust the
delay as it's 3 seconds by default making the time we check longer
than the async tasks. Eventually, we add a new parameter for
controlling the downloading of the image
"OVNControllerImageUpdateTimeout" which is set to 10min by default.

Backport conflicts:
- Testing ci is different and we don't need that part of the patch;
- podman_image instead of full containers.podman.podman_image as this
  isn't supported in train (module not found)
- Some code from wallaby aren't backported, so by default code
  conflict occured: one of them is the excluded_overcloud code that will
  find its way later in train.

Change-Id: I6d439367816d5f5968b8f890f74b7a2f74715f52
(cherry picked from commit 76dec8bc6ad9a07ffc497a050330931d5e38dfcf)