Restrict networks used by nova migration target
Currently the sshd service for nova migration target listens to all
networks and the match block is used to restrict the networks allowed
for the nova_migration user.

This was required when the host sshd service was used, but now the nova
migration target service is independent from the host sshd and we can
just restrict the networks used by the services.

Depends-on: https://review.opendev.org/842858
Change-Id: I506b157c0de7232181a2a9c22bcf632fa67a8d0d