Keystone: Remove unnecessary binds for admin endpoint
In the past Keystone required a separate instance for admin endpoint.
However that requirement was removed and current TripleO deploys only
a single keystone instance, which listens on tcp/5000 in
KeystonePublicApiNetwork [1].

Now haproxy forwards all keystone api access to tcp/5000 in
KeystonePublicApiNetwork and no longer uses backend on tcp/35357 in
KeystoneAdminApiNetwork.

[1] 5bd82e9ef051e732c0f991eaf4f9e3572aac87f6

Closes-Bug: #1978131
Change-Id: I9daaf69463e74f0f7fb63586ff6464194391da30