Restore missing SAN IP in etcd cert
This patch fixes a regression in the etcd cert that is used with
tls-everywhere. The etcd service makes cluster connections using
its IP address, so with tls-e the etcd cert needs to include its
address as a SAN IP. The SAN IP was inadvertently dropped when
cert generation was moved from puppet to an ansible role.

Resolves: rhbz#2096230
Change-Id: I28e268c58bf63aaa81a67f48ca492fe97d9a07e9