Swift: Do not enable the whole rsync::server class
In puppet run for swift, we enable the entire rsync::server class and
this sometimes results in unexpected package installation inside
the container-puppet-swift container.

Because we only need file/concat resources defined in the class since
we get rid of xinetd for CentOS 9 and these resources are enabled by
default, this change removes that entire class from tags.

Related-Bug: #1979204
Change-Id: I7e29a6e797155ca5c5aa27996a4e35e688061b6f