Fix nova_libvirt_max_queues_set conditions for nova-compute
A typical pattern to check for a feature gate by a param and its role
param (there are more types of checking, this one is for negation):
foo_gate_enabled:
  not:
    and:
      - equals: [{get_param: foo_val}, <null_val>]
      - equals: [{get_param: [RoleParameters, foo_val]}, <null_val>]

which is consider the feature gate OFF, only if both of its tht param
and role params are holding empty/falsey (null) values. If either holds
a not-null value, consider the condition is True, and treat the feature
gate as enabled.

For the nova compute service template, we follow that pattern when
checking permit_auto_converge_set, permit_post_copy_set,
reserved_huge_pages_set.

But we do not, when we check it for libvirt_file_backed_memory_enabled
and nova_libvirt_max_queues_set. And consider the feature disabled, if
*either of* the tht, or role param holds a null value. As a result,
when we define NovaLibvirtMaxQueues, or NovaLibvirtFileBackedMemory,
but have nothing for those in roles params, the features will not
become enabled.

Related: rhbz#1917460
Related: rhbz#1922819
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: Ie16ab2ab5836ed455aeda35d2ba619633a595411