Fix Cinder Dell EMC SC backend template
The 'CinderSCStorageProtocol' parameter defined in
cinder-backend-dellemc-sc-puppet.yaml is not used and
has no effect. Modified it as 'CinderScStorageProtocol'
similar to other parameters.

The parameter 'CinderScExcludedDomainIp' is introduced
when the backend support was added in the patch
https://review.opendev.org/c/openstack/tripleo-heat-templates/+/727283
but it is never used.
Removed it as the 'CinderScExcludedDomainIps' parameter is already used.

Closes-Bug: #1978749
Closes-Bug: #1978750
Change-Id: I6a82a542855e32bd82c67f34138271b060df5de2
(cherry picked from commit 95a02d2419fa279838a5bc3b600b235f41615cac)