Remove leftover of N1kv driver support
We used to use the mechanism_drivers parameter in horizon so that we
can enable the cisco profile support automatically when N1kv driver is
enabled.

However N1kv driver support was already removed by [1] and the logic
was already removed from puppet-tripleo. So that puppet parameter
injection is no longer used now.

[1] 8310c55b08e5d9dbe6d99ca1662113961acf6aef

Change-Id: I2d10f409fb537a9f48a35cd225342646ed3d3e1a