Allow for jinja2 templating of docker_config
In support of the wider Multi-RHEL effort. We will
need to support templating to allow for Ansible to
determine which container needs to be pulled to
the hosts based on their OS version.

This change adds support for that templating of
the docker_config files.

Change-Id: Id56a4670507524f41fde238887fce4dd19e90a87
(cherry picked from commit 3d40e77a66eb7121a20d59c65788c89fccbd2ca2)