Set OVSDB Connection.probe_interval
With ovn_cluster, connections were set with --remote on the CLI by
ovn-ctl (except for a workaround for TLS connections). Via this
method, there is no way to set the inactivity probe interval which
is a regression from using non-clustered OVSDB.

To achieve setting this value, the connection must be defined in
the OVN Connection table. For RAFT, since each server shares the
same DB, this means using a single Connection entry that all
servers use, which necessitates listening on 0.0.0.0. To make this
safe, the firewall_rules are adjusted to only allow OVSDB traffic
from the proper subnet.

A subsequent patch will be made to tripleo-ansible that removes
passing the --create-insecure-remote options to ovn-ctl which,
for non-TLS, set ovsdb-server to listen on its specific address.

Resolves: rhbz#2110550
Related-Bug: rhbz#2099805
Related-Bug: rhbz#2101588
Related-Bug: rhbz#1952038
Change-Id: I7d8d0530c367708215437c9ac11a6fc17235e784