Support role specific ContainerImagePrepare
This implements partial support for overriding the ContainerImaePrepare
parameter for individual roles, so that users can use different set
of images for each role.

example:
parameter_defaults:
  ContainerImagePrepare:
    - push_destination: true
      set:
        name_prefix: openstack-
        name_suffix: ''
        namespace: quay.io/repository/tripleowallabycentos9
        ...
  ComputeEl8ContainerImagePrepare:
    - push_destination: true
      set:
        name_prefix: openstack-
        name_suffix: ''
        namespace: quay.io/repository/tripleowallabycentos8
        ...

This change ensures that container image prepare process, which is
executed by external tasks, pulls container images according to these
layer definitions so that the required container images are downloaded
and pushed to the undercloud image-server during deployment.

Depends-on: https://review.opendev.org/c/openstack/tripleo-ansible/+/851689
Change-Id: I7c96d7d0f23fd999e76b417f456ac6a2764c4a3e