Use python to template cell urls
Heat and Puppet cannot handle it properly due to url_parse()
not escaping url encoded special characters.

In future, this script could be reworked into an ansible plugin.
We need it as is to close the related multi-cells regression
sneaked in after Train for Wallaby.

This doesn't raise the bar for standalone roles migration plans
as we have already a plenty of container config scripts to be
migrated as well. Adding another one won't make things (much)
worse.

Related: rhbz#2089512
Change-Id: I0216dd555eacbcb37abd03a67fdd39f9244285db