Change --start-at-task for plays using tripleo_free
With tripleo_free strategy a playbook run fails only after the
end of the play. In case a task failure for a node, the other nodes
keep continuing till the end of play. Therefore, --start-at-task
should be always at the beginning of the play as different nodes
can fail at different tasks and we need to always start from the
start of play.

Change-Id: I79ff5d3babd05c7d1ba5195f372c1348d6692e2b