Fix TRIPLEO_CONFIG_HASH generation
There are times when there is no newline char at the end of files
that are used to a generate config hash. This is will result in
header lines with datetimestamp getting appended to the previous
line and not filtered. So, container-puppet.py would generate a
different TRIPLEO_CONFIG_HASH for services every time.

Ex. When using custom policies it would append puppet generated line
from cron config with timestamp to the last line from policy.json.
This would result in new config hash for every deployment and
containers getting replaced.

Resolves: rhbz#2083016
Change-Id: I2153844ed06756eeb9ff56cfff64908bed5fa577