Deprecate snmp service
The snmp service was first introduced so that undercloud ceilometer can
gather metrics via the snmpd daemon in overcloud nodes.

However Telemetry services are already removed from undercloud, and
this requirement is no longer valid.

This deprecates the snmp service support so that we can remove it in
a future release. This also introduces upgrade tasks to clean up
the service because it has been enabled by default.

Note that the net-snmp package will be removed from the overcloud image,
and using the disabled snmpd service requires adding snmpd package into
the overcloud image or setting up a package repository.

Change-Id: I9d13fccb5d4d265cfc04df5ccb287b6215e7d0b2