[train-only] Don't update hash when using paunch
In train with paunch we use hased-<service>.json files for
container startup configs. However, with ansible we use
<service>.json in later versions. The task to update
config hash in step_1 deletes the hashed<service>.json  file
and creates <service.json> from old configs. This creates
lot of of confusion as we end up with two different files
with old and new config hashes. Paunch uses hashed-<service>.json
files. We should skip the task to update config hash when paunch
is enabled and remove the config when a hashed file is created.

Change-Id: I6dc8c4bb64e7b28266337f067b25985e82aec42c