[Stable only] Apply IronicIPXEUefiSnpOnly to uefi_pxe_bootfile_name
Currently, uefi_pxe_bootfile_name will be set to bootx64.efi but
bootx64.efi does not exist because a change[1] of puppet-ironic is not
backported. In this situation, deployment will fail. Ideally, the
change[1] of puppet-ironic should be backported, but it is not going to
be backported.

Therefore, this patch changes uefi_pxe_bootfile_name to snponly.efi or
ipxe.efi in response to IronicIPXEUefiSnpOnly.

In master branch, the change[1] of puppet-ironic has been made and
boot64.efi is created. Hence this change is only for Stable branches.

[1]4863a485cd7405f6bca9011a9694dd385d170049

Closes-Bug: #1983559
Change-Id: I0eb9aca944479cd242468aa1674555edcefba56b