[Stable only] Apply IronicIPXEUefiSnpOnly to uefi_pxe_bootfile_name
Currently, uefi_pxe_bootfile_name will be set to bootx64.efi but
bootx64.efi does not exist because a change[1] of puppet-ironic is not
backported. In this situation, deployment will fail. Ideally, the
change[1] of puppet-ironic should be backported, but it is not going to
be backported.

Note for Train:
  On Train, uefi_pxe_bootfile_name will be set to ipxe.efi regardless of
  IronicIPXEUefiSnpOnly because a change[2] was not backported to Train.

Therefore, this patch changes uefi_pxe_bootfile_name to snponly.efi or
ipxe.efi in response to IronicIPXEUefiSnpOnly.

In master branch, the change[1] of puppet-ironic has been made and
boot64.efi is created. Hence this change is only for Stable branches.

[1]4863a485cd7405f6bca9011a9694dd385d170049
[2]bd7c489dc7346f071f15080c2b6c4715e39bbb3c

(from stable/wallaby to stable/train)
Conflicts:
	deployment/ironic/ironic-conductor-container-puppet.yaml

Closes-Bug: #1983559
Change-Id: I0eb9aca944479cd242468aa1674555edcefba56b
(cherry picked from commit 6435d081b2b6a30023e2dfa27666799ec55ddbfc)