Introduce ExternalCeph boolean
Add explicit parameter for when external ceph is used.
This parameter defaults to false but is true if the
deployment uses -e environments/external-ceph.yaml.

When external ceph is used with Ganesha, the ceph_mon
group is empty but the ceph_nfs group is not. When
internal ceph is used with Ganesha, both the ceph_mon
and ceph_nfs groups are non-empty. Rather than solve
the related bug by adding another condition based on
these groups  which is compatible existing logic, it's
safer to have an explicit parameter for when external
ceph is used.

Change-Id: Id3e397d81dbca9a48d0456588784bcc20737093f
Depends-On: I2651c2850debd8110da93df2adc5fd8768a00db0
Closes-Bug: #1986988