Allow for jinja2 templating of puppet_config
In support of the wider Multi-RHEL effort. We will
need to support templating to allow for Ansible to
determine which container needs to be pulled to
the hosts based on their OS version.

This change adds support for that templating of
the puppet_config files.

Change-Id: I5b6a34c4013ad86585bdbd29dab882dd0dd1b3e2
(cherry picked from commit 524242606f5ce17c1002aa8f45c4a4dce7bc9c20)