Remove redundant override of vhost tcp port
The latest puppetlabs-apache release (8.0.0) introduced stricter type
validations and a string value is no longer accepted for the
apache::vhost::port parameter. Because of this change we are no longer
able to use string values for *::wsgi::*::api_port parameters.

Because the values currently defined in templates are exactly same as
the default values defined in underlying puppet modules (puppet-nova
and puppet-placement), we can just remove these overrides and use
the puppet default values, which are compliant with the new type
validations.

Closes-Bug: #1987429
Change-Id: I48ad8babf694634bc1107402fd729bdc39fc109e