Use NeutronInterfaceDriver parameter consistently
Currently the NeutronInterfaceDriver parameter is only partially used
and the parameter is sometimes ignored because of the hieradata defined
in individual templates. This is quite confusing and causes unexpected
results.

This change ensures the parameter is consistently used to manage
the interface_driver parameter of dhcp agent and l3 agent.

Change-Id: I9955a8493bb031674a9d99087bd83b009011a0a2