Add role specific settings for ovn_bgp_agent parameters
Before this patch only the agent driver could be configured
per role.

With this change we also allow to configure:
- the expose_tenant_networks flag, as this only need to be enabled
on the networker nodes
- the AS number, in case this is different between different nodes
- the ovsdb_connection

In addition it adds an extra parameter that was missing, the
reconcileInterval that allows to configure how frequently the
status is reconciled. The reason to make it role specific is because
it may be the case that different intervals are preferred for
networker and compute nodes, as the number of events to manage
may be different.

Change-Id: Ieaee6e162b796853445b3ecc941c61e2411d3ada