Redis: replace hiera CLI
The hiera CLI is deprecated[1] in favor of the new ``puppet lookup``
command and doesn't work with the latest hieradata v5. However
the ``puppet lookup`` command requires gathering fact data and should
not be used outside of configuration tasks.
 [1] https://puppet.com/docs/puppet/7/hiera_automatic.html

This replaces the hiera CLI used to detect the listen address of
the stunnel service for redis, by parsing the stunnel.conf.

Change-Id: Ib1ed9bb4ba1e2d40ed415ddaa43a8528a3d058a9