Correct label for /run/libvirt
The former label (virt_var_run_t) was in fact the one set by fcontext:
/var/run/libvirt(/.*)?                             all files          system_u:object_r:virt_var_run_t:s0

While it makes sense when libvirt is running on the host, it has no
actual value in containerized services.

More over, it was already relabeled to container_file_t from within the
migration-target container, since that one bind-mounts the location with
the "z" flag.

In order to stop seeing the label flapping upon deploy and day-2
operations, the best thing to do is to stop using the virt_var_run_t
label and just ensure containers are relabeling this location.

Resolves: rhbz#2122656
Change-Id: I64f7e5d5f7dab8e59c6a48f01d636880e429d2f2