Move podman socket
This patch creates special podman socket for sensubility purposes
on /var/lib instead of /run as it is tmpfs and so permission changes
would perish after reboot.

This patch also removes rsyslog configuration for separate HC log file.
This is not needed anymore since sensubility fetches data straight from podman
instead of parsing the log file.

Sensubility is also set to connect to QDR in scenario001-standalone job
so that it is visible in the logs that HC script succeeds.

[1] https://github.com/containers/podman/issues/12493
Depends-On: I7538def60e4fd3a7187ede0e1f5ada1b0c460f65
Change-Id: I9149dd5087ae6af0265f37b36de9213c9d625a3f