Enable dns_domain_ports extension by default
'dns_domain_ports' extension allows dns_domain to be set
for ports and ML2 OVN supports it since wallaby[1]
and ML2 OVS long before that.
This patch replaces 'dns' with 'dns_domain_ports' for
NeutronPluginExtensions in ml2 plugin heat template and
all environment files.
'dns' extension not equired to be enabled explicitly
with 'dns_domain_ports' as it inherits all the capabilities
of 'dns' extension.

[1] https://review.opendev.org/q/Ic269902ef7a16893c4ea624b04347d04db4f52d9

Related-Bug: #1990480
Change-Id: I25f75d7bd8864ea132246cea79e073232d664d76