Fix inconsistent <service>Workers
Currently the <service>Workers parameters are implemented in different
ways according to services. Because of this, these parameters enforce
different type(number vs string) and behave differently for a specific
value such as 0. These are quite confusing for users.

This change introduces the standard implementation for all Workers
parameters so that the parameters behave consistently in all services.

This also introduces the following changes as side-effects.

- Outdated description of some workers parameter is removed. Currently
  workers are limited to a certain number(such as 12) even in a system
  with many CPU cores

- Default value of BarbicanWorkers is effectively lowered down to
  $::os_workers from $::processorcount, to use the consistent default
  workers for all services.

Change-Id: Ie7a12a297e00efebca017d51afb9bf9e8c56ee50