Set external_ids:ovn-ofctrl-wait-before-clear
In order to avoid flow update while updating, we need to set
external_ids:ovn-ofctrl-wait-before-clear to some value that will
be used to wait before clearing all flows.

There is code for taking this parameter into account during
deployment, this is the part that takes it into account during update.

Change-Id: I8af36a960156e5203c97293442fbfe1a1ff47dd5
Co-Authored-By: David Hill <davidchill@hotmail.com>