Run nova_wait_for_compute_service only for the default cell deployment
nova_wait_for_compute_service should not run when deploying an
additional cell. For the additional cell scenario we can not run
the discovery automatically as the cell first needs to be created.
Therefore we can also not wait on computes to be registered and
queried via the nova api.

The issue is here [1] where it runs the nova_wait_for_compute_service
if `{get_param: NovaAdditionalCell}`

            - if:
                - {get_param: NovaAdditionalCell}
                - nova_wait_for_compute_service:

NovaAdditionalCell defaults to false as NovaAdditionalCell is the
special case. In train this is if `is_not_additional_cell`

            - if:
                - is_not_additional_cell
                - nova_wait_for_compute_service:

The simplification patch in [3] introduced this difference.

So in general it seems that nova_wait_for_compute_service is right
now only running when deploying an additional cell, which it
should not.

[1] https://github.com/openstack/tripleo-heat-templates/blob/stable/wallaby/deployment/nova/nova-compute-container-puppet.yaml#L1412
[2] https://github.com/openstack/tripleo-heat-templates/blob/stable/train/deployment/nova/nova-compute-container-puppet.yaml#L963
[3] https://github.com/openstack/tripleo-heat-templates/commit/476e444b8f23488eae2c0fd712bbbb30f34ec5af

Change-Id: I0401a9a28d4715073fb929018dfb6daabe836cd8
(cherry picked from commit 38dbd331cccb0c47739462ca8261127863b41054)