Replace dnf by tripleo_dnf_stream for undercloud upgrade
Dnf Ansible module doesn't only enable the stream, but
synchronizes the whole content. Dnf module installs
packages which were not installed before the undercloud upgrade.
As a result some packages are installed in step0, and not as it
should be done in step3.

This patch replaces the dnf Ansible module by a custom tripleo
module, tripleo_dnf_stream, which will only enable the stream
(if the stream wasn't enabled).

This is the same change that was made for minor update workflow,
but for undercloud upgrade. [1]

[1]https://review.opendev.org/#/q/I3fa4ccc224cf510bcc85a9c86a2b4f0d367c687f

Change-Id: I3c24f6c0b1877caca225d532e2c40cabe9095577