Add legacy monolithic libvirt service for FFUv3
FFUv3 (Train->Wallaby) and mixed-rhel stories require possibility to
run modular libvirt on el9 hypervisors and monolithic libvirt on el8.

It is not clear yet if we might need a new ComputeEL8 role, or not.
Create legacy service OS::TripleO::Services::NovaLibvirtLegacy to run
monolithic libvirt on such computes that are supposed to stay on
RHEL8.4 (el8) for some time.

Related: rhbz#2136163
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: I0476851066a5c3e22e09ff02ed4bb19a7eff71a7