Fix ovn_controller certificate principal
https://review.opendev.org/c/openstack/tripleo-heat-templates/+/822313/
changed the way the certificate principal is defined in hieradata to
match what is returned by getcert list, so that the puppet comparisons
would match correctly.  Otherwise, puppet would think that the cert
resource has changed and always request a new cert.

One of the certificates was missed though - ovn_controller.

Change-Id: I53230f0548245c613a2f8cbff9c3c51259c35f59