[TRAIN-ONLY] Collect and parse all health check logs
Previously some of the container health check run log records were not collected
by rsyslog and so health status of some containers was not reported.

This patch makes it so all relevant log records from systemd-timer based health
check runs are collected and correctly parsed and so those containers with health
check attached will be reported.

Wallaby+ is using podman socket for fetching container status data, so this patch
is needed only for stable/train.

Resolves: rhbz#2149002

Change-Id: Ie0dc9edac5dc1483c9d419a169457a8689be4064