[TRAIN-ONLY] Keep the cephx key out of /var/log/messages
When creating nova's libvirt secret, look up the cephx key from the
client keyring instead of passing it directly on the command line.
The command line ends up in the system log, so dynamically looking
up the cephx key prevents it from appearing in /var/log/messages.

This is train-only because the corresponding code in wallaby already
avoids passing the cephx key on the command line.

Change-Id: Ibaec4febdd19508f799134da3662e4e418c096a4