[ovn] Don't add conntrack entries for Geneve
As Geneve UDP traffic is allowed, there's no reason to create
conntrack entries as it may result in a performance hit.

This patch is preventing Geneve traffic to be sent to conntrack.

Fix for Train only:
firewall/rule.pp add NEW as default state in the rules,
this fix add INVALID as state for the geneve UDP rules

Closes-Bug: #1885551
Change-Id: I1b582b86780881fe7228540edf281690733fb946