Add Redis password to designate central service
Make sure that at the redis password is available for configuring
process coordination in the event that the central and producer
services are spread out over separate servers via composable roles.
This was missed when the related change
Icc23208945304913cdaeede703c486064646379a was proposed.

Note: this patch should be backported through to wallaby for the
related puppet-tripleo patch already merged.

Change-Id: Icef9c5cbed256d977590f2e38427a1155b503824