Replace qdrouterd backend with rabbitmq in Sc03
qdrouterd for RPC is not a supported backend in RHOSP,
updating sc03 to test rabbitmq which is a supported
backend for RPC communication.

Trying to run with qdrouterd backend in downstream fails
because python-pyngus is not added as dependency downstream[1]
python-pyngus available since RHOSP13 but is not added as dep
for oslo-messaging rpm with reasoning[2].

~~~
 ModuleNotFoundError: No module named 'pyngus'
~~~

Sc03 provides desginate testing coverage, This change will
allow us to run standalone sc03 in downstream pipelines keeping
sc03 env files in sync in upstream and downstream.

This patch will be backported to wallaby.

[1] https://github.com/rdo-packages/oslo-messaging-distgit/blob/wallaby-rdo/python-oslo-messaging.spec#L89-L91
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1791364#c13

Change-Id: Ib49223709e20931abcec4d16301db9882c64c06f
(cherry picked from commit 9962b52a0a2ff93d57e9d1dff8a03d70991df1a8)