Set ovn-monitor-all before ovn-controller update.
This parameter is missing during ovn-controller update and can cause
cut in data plane in certain cases.

We refactor this section of the external_update_tasks to manage all
parameters that have to be set during update for ovn-controller. If
others are needed we just add them to that section.

Change-Id: I30f083a906e457fd5b92d96a6fda8e6451c11fbb