Add update of OVN cluster container.
Currently OVN cluster images are not updated during update.

For container to be updated we rely on either:
- explicit update tasks: like it's done for pacemaker for instance;
- docker_config tasks: those triggered by common_deploy_steps_tasks;

Here the ovn cluster is setup using ansible and no extra update tasks
are present, which makes the update process to ignore change in the
container image configuration.

We add the necessary code block in the update tasks so that the
cluster container are updated using anchor and alias to avoid
repetition.

Furthermore we move the generation of the new container definition (in
/var/lib/tripleo-config/container-startup-config) earlier in the
update process. Doing so enable the new image definition to be
available during the update tasks.

Closes-Bug: #2004575
Change-Id: Ieeea6267b7c35a76f83a463a7247802cba99c4b5