Revert "Run virtqemud with umask 0027"
This reverts commit 7bba86fc582b6d599117be06c23d0a3d91332cfd.

Reason for revert: Fix regression for virtqemud is started with umask of 0027 which then gets in the way libvirt sets up hugepages. In this particular case, libvirtd/virtqemud creates $hugetlbfs/libvirt/qemu/ (/dev/hugepages/libvirt/qemu/ in this case) which is meant as top level directory, a base where individual, per-guest directories are then created. Now, the top level directory is owned by root:root (so that nobody else can modify its perms) but we also want it to be accessible by other processes (mind you, we don't need to list the directory (o+r), we need to be able to access subdirs (o+x)).

Setting such restrictive umask for a process that is supposed to prepare environment for other users (QEMU runs under qemu:qemu) was a bad choice because of that, hence this revert

Change-Id: I75897fbaf7d9415178bc267c848b83775a384814
Related: rhbz#2166979