Remove neutron sidecar containers in more reliable way
Previously kill-script used by Neutron agents to kill sidecar
containers was sending TERM signal to container when it was going to be
closed with SIGTERM. After that it was immediatelly trying to remove
container with "podman rm" command.
Sometimes, if container wasn't really stopped yet when "podman rm" was
used, removal of container failed and there was leftover sidecar on the
host.
Now kill-script will instead use "podman stop" command which will send
TERM to the container and wait some time (10 seconds by default) until
it will be really stopped before it will try to delete container.

Additionally this patch improves logging in different functions in
kill-script.

Closes: rhbz#2080811
Change-Id: If893bb6883843bea48c50c47b98f182cacdd50b8