Fix cinder-backup when running active/active
Historically, cinder-backup runs as root under pacemaker, and it's
necessary for it to run as root when running active/active (i.e.
not under pcmk). This is necessary for two reasons:

First, os-brick lockfile permission conflicts arise if c-vol runs
as root (under pcmk) but c-bak doesn't run as root. c-bak won't have
permission to create the lock files when it uses brick to attach to
the volume it's backing up or restoring.

The second reason pertains to using c-bak's NFS backend. Cloud admins
typically provide an NFS share that allows write access to root
clients, but not to other clients such as 'cinder' (in fact, the
'cinder' UID will probably not even be known on the NFS server).

Change-Id: I343fb17fb1aed695a8355b292b56157c70c3ed31