Allow 1993 port for HAProxy stats page
HAProxy listens on 1993/tcp port for its stats page by default
However, currently, 1993 port is not allowed when
deployment/haproxy/haproxy-pacemaker-puppet.yaml is used because that
file is not inherits firewall_rules parameter from
haproxy-container-puppet.yaml.
By this change, firewall_rules is inherited and 1993 port will be
allowed when haproxy-pacemaker-puppet.yaml is used.

Resolves: rhbz#2187877
Change-Id: Iab1fa49ccc1244b03535c11fd1e1e6899600bba7