Support update of ovn-controller when deployed in step4.
A recent review[1] change the step in which ovn-controller was
deployed. It was step 4 before and now it's step 3. This breaks the
ovn-controller update for environment deployed with step 4[2]. The net
result was un-updated ovn-controller during the ovn-controller
update. This eventually leads to cut in data plane when the ovn
cluster is updated.

This patch introduces support for both steps by testing if the step_4
file exists on one of the host belonging to the ovn_controller group.

We test for empty group in the delegate_to task mainly to pass ci as
in a normal deployment this should always be true as the template is
included.

[1] https://review.opendev.org/c/openstack/tripleo-heat-templates/+/879519
[2] In downstream term that would be 17.0

Change-Id: Icd5faed6334bf9b06f5526a4f0cb62a2f28cbca1